YESS ANNOTATION OF OTHER PERSON WORKS!
well, the red pin showed up! just going to commit this so I don’t lose
it. viewForAnnotation was going into an infinite loop and not producing
anything and it turned out to be because I was called [mapview
setseelcted:annotation …]. Not sure why this was an issue, but alright.